requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
boto3==1.34.0
Pillow==10.4.0
PyPDF2==3.0.1
pdf2image==1.17.0
gunicorn==21.2.0
flask-cors==4.0.0

